Moved login package to presentation package
