+++ app/src/main/res/feature_auth/common/navigation/login_navgraph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/loginFragment"
-        android:name="com.hopcape.findmy.feature_auth.login.LoginFragment"
+        android:name="com.hopcape.findmy.feature_auth.presentation.login.LoginFragment"
         android:label="fragment_login"
         tools:layout="@layout/fragment_login" >
         <action
